Fixed animations for enchantment icons and inventory items.
- Passed in new delta time into ModuleEvent() as a double to fit va_list and prevent data corruption.
- Removed old dt from animate.c
- Updated function parameter for animating inventory item and enchantment icons to use float.

+++ module/merintr/enchant.h
@@ -25,7 +25,7 @@ bool EnchantmentDrawItem(HWND hwnd, const DRAWITEMSTRUCT *lpdis);
 void EnchantmentsResize(int xsize, int ysize, AREA *view);
 void EnchantmentsRedraw(void);
 void EnchantmentsNewRoom(void);
-void AnimateEnchantments(int dt);
+void AnimateEnchantments(float dt);
 
 #endif /* #ifndef _ENCHANT_H */
